week2/day2_speech_ai/requirements.txt
openai>=1.30.0
deepgram-sdk>=3.2.0
elevenlabs>=1.2.0
sounddevice>=0.4.6
soundfile>=0.12.1
httpx>=0.27.0
websockets>=12.0
python-dotenv>=1.0.0
numpy>=1.26.0
